requirements.txt
Flask
pandas
numpy
seaborn
scikit-learn
matplotlib
requests
cassandra-driver
-e .